refactor(@sounisi5011/encrypted-archive): rename some files in `src`
+ `src/core/utils/stream.ts` -> `src/core/utils/reader.ts`
+ `src/core/stream.ts`       -> `src/core/utils/convert.ts`
+ `src/core/utils/type.ts`   -> `src/core/types/utils.ts`
