+++ packages/encrypted-archive/src/core/header/protocol-buffers-converter/utils.ts
@@ -1,6 +1,6 @@
 import type { BuiltinInspectRecord } from '../../types/builtin';
+import type { Cond2, Nullable, OneOrMoreReadonlyArray } from '../../types/utils';
 import { getPropFromValue } from '../../utils';
-import type { Cond2, Nullable, OneOrMoreReadonlyArray } from '../../utils/type';
 
 function reportNonDefinedField(opts: { fieldName: string; dataName: string }): never {
     throw new Error(`${opts.fieldName} field in ${opts.dataName} is not defined`);
