+++ packages/encrypted-archive/src/core/key-derivation-function/argon2.ts
@@ -10,8 +10,8 @@ import {
     GetArgon2KDFResult,
     NormalizedArgon2Options,
 } from '../types/key-derivation-function/argon2';
+import { assertType, isInteger, objectEntries, objectFromEntries, RequiredExcludeUndefined } from '../types/utils';
 import { ifFuncThenExec, isNotUndefined, normalizeOptions, passThroughString } from '../utils';
-import { assertType, isInteger, objectEntries, objectFromEntries, RequiredExcludeUndefined } from '../utils/type';
 
 export const defaultOptions: NormalizedArgon2Options = {
     algorithm: 'argon2d',
