+++ packages/encrypted-archive/src/core/types/key-derivation-function/argon2.ts
@@ -1,5 +1,5 @@
-import type { RequiredExcludeUndefined } from '../../utils/type';
 import type { BaseKeyDerivationOptions, GetKDFResult } from '../key-derivation-function';
+import type { RequiredExcludeUndefined } from '../utils';
 
 export const argon2TypeNameList = ['argon2d', 'argon2id'] as const;
