+++ packages/encrypted-archive/src/core/types/utils.ts
