+++ packages/encrypted-archive/src/index.node.ts
@@ -16,8 +16,8 @@ import type {
 } from './core/types';
 import type { CryptoAlgorithmName } from './core/types/crypto';
 import type { KeyDerivationOptions } from './core/types/key-derivation-function';
+import type { Expand, ExpandObject } from './core/types/utils';
 import { convertIterableValue } from './core/utils';
-import type { Expand, ExpandObject } from './core/utils/type';
 import { getCryptoAlgorithm } from './runtimes/node/cipher';
 import { CompressOptions, createCompressor, decompressIterable } from './runtimes/node/compress';
 import { kdfBuiltinRecord as kdfBuiltin } from './runtimes/node/key-derivation-function';
