+++ packages/encrypted-archive/tests/helpers/index.ts
@@ -1,6 +1,6 @@
 import { TextEncoder } from 'util';
 
-import type { AsyncIterableReturn } from '../../src/core/utils/type';
+import type { AsyncIterableReturn } from '../../src/core/types/utils';
 
 export function isOneOrMoreArray<T>(value: T[]): value is [T, ...T[]];
 export function isOneOrMoreArray<T>(value: readonly T[]): value is readonly [T, ...T[]];
